% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEmrt.R
\name{FEmrt}
\alias{FEmrt}
\title{Fixed effects meta-regression tree}
\usage{
FEmrt(formula, vi, data, c = 1)
}
\arguments{
\item{formula:}{A formula to specify the response variable (usually the effect size) and the interested moderators. It should be written like y ~ x1 + x2.}

\item{vi:}{The name of the column of the sampling variance of each study in the data set.}

\item{data:}{A meta-analytic data set. Needs to be a data frame.}

\item{c:}{The pruning parameter to prune the initial regression tree by "c-standard-error" rule. Needs to be a non-negative scalar.}
}
\value{
If no moderator effect is detected, the function will perform a standard meta-analysis and return a list including the following objects:

no.: The total number of the studies

Q: The Q-statistics for the heterogeneity test

df: The degree of freedoms of the heterogeneity test

pval.Q: The p-value of the heterogeneity test

g: The overall effect size for all studies, based on Hedges'g

se: The standard error of the overall effect size

zval: The Z value of the z-test

pval: The p-value of the significance test for the overall effect size

ci.lb: The lower bound of the confidence interval for the overall effect size

ci.ub: The upper bound of the confidence interval for the overall effect size

formula: The formula that was specified in the model

If  moderator effect(s) is(are) detected, the function will perform a subgroup meta-analysis and return a list including the following objects:

tree: The tree that represents the moderator effects and interaction effects between the moderators. An rpart object.

labs: A data frame showing how the studies were split by the moderators into subgroups

no.: The number of the studies in each subgroup

Qb: The between-subgroups Q-statistics

df: The degree of freedoms of the test for between-subgroups heterogeneity

pval.Qb: The p-value of the test for between-subgroups heterogeneity

Qw: The within-subgroup Q-statistic in each subgroup

g: The overall effect size in each subgroup, based on Hedges'g

se: The standard error of each overall effect size

zval: The Z value of the z-test for each overall effect size

pval: The p-value of the significance test for the overall effect size of each group

ci.lb: The lower bounds of the confidence intervals

ci.ub: The upper bounds of the confidence intervals

formula: The formula that was specified in the model
}
\description{
Apply fixed effects meta-regression trees to meta-analytic data sets.
Note that the model is assuming fixed effects within and between subgroups.
In other words, it is assumed that no residual heterogeneity is present.
}
\examples{
data(SimData)
test <- FEmrt(efk~m1+m2+m3+m4+m5, vark, data=SimData, c=0)
test
plot(test)
}

