% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEmrt.R
\name{plot.FEmrt}
\alias{plot.FEmrt}
\title{Visualisation of Fixed Effects Meta-Regression Tree}
\usage{
## S3 method for class "FEmrt"
plot(x, type=4)
}
\arguments{
\item{x:}{A FEmrt object.}
}
\description{
Plot function for an FEmrt object. The plot function uses the plot method from the package rpart.plot
 of Stephen Milborrow. The plots shows the result of fixed effects meta-regression tree with subgroups
 split by moderators and the corresponding estimated overall effect sizes.
}

