% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.metacart}
\alias{predict.metacart}
\title{Predictions from a fitted metacart object}
\usage{
## S3 method for class "metacart"
predict(x, newdata)
}
\arguments{
\item{x:}{fitted model object of class "metacart".}

\item{newdata:}{data frame containing the values at which predictions are required.}
}
\value{
A data frame containing the predicted effect size, the moderators, and the corresponding node lables in the fitted tree.
}
\description{
Returns a data frame of predicted effect sizes and moderators from a fitted metacart object
}

