% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEmrt.R
\name{print.FEmrt}
\alias{print.FEmrt}
\title{Pinrt function for FEmrt}
\usage{
## S3 method for class "FEmrt"
print(x, digits = 3)
}
\description{
Print out the results of an FEmrt object
}
\details{
If no moderator effect is detected,
the print function will show the standard meta-analysis results.
Otherwise, the print function will show the subgroup meta-analysis results,
with the significance test resutls for moderator effects, the splitting points of the moderators,
and the estimated effect sizes in all subgroups.
}

