% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintSummaryPlot.R
\name{print.metacart}
\alias{print.metacart}
\title{Pinrt function for metacart}
\usage{
## S3 method for class "metacart"
print(x, digits = 3)
}
\description{
Print out the results of an metacart object
}
\details{
If no moderator effect is detected,
the print function will show the standard meta-analysis results.
Otherwise, the print function will show the subgroup meta-analysis results,
with the significance test resutls for moderator effects, the splitting points of the moderators,
and the estimated subgroup overall effect sizes.
}

