% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintSummaryPlot.R
\name{summary.metacart}
\alias{summary.metacart}
\title{Summary function for metacart}
\usage{
## S3 method for class "metacart"
print(x, digits = 3)
}
\description{
Summary the results of an metacart object
}
\details{
If no moderator effect is detected,
the summary function will show the standard meta-analysis results.
Otherwise, the summary function will show the subgroup meta-analysis results,
with the significance test resutls for moderator effects, the splitting points of the moderators,
and the estimated subgroup overall effect sizes.
}

