% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEmrt.R
\name{treepruner}
\alias{treepruner}
\title{Prune a tree}
\usage{
treepruner(tree, c)
}
\arguments{
\item{tree:}{A initial tree fitted by rpart, needs to an rpart object.}

\item{c:}{A scalar to prune the  tree by selecting the tree with minum cross-validation error plus the standard error multiplied by c.}
}
\value{
The pruned tree, also an rpart object.
}
\description{
Prune an initial rpart tree by "c-standard-error" rule.
}

