% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintSummaryPlot.R
\name{print.metacart}
\alias{print.metacart}
\title{Print function for metacart}
\usage{
## S3 method for class "metacart"
print(x)
}
\description{
Print the results of an metacart object
}
\details{
The function returns the objects concerning the analysis results.
}

