% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintSummaryPlot.R
\name{summary.metacart}
\alias{summary.metacart}
\title{Summary function for metacart}
\usage{
## S3 method for class "metacart"
summary(x)
}
\description{
Summary the results of an metacart object
}
\details{
The function gives summary of the analysis results and details of all return values.
}

